monitor-bot/requirements.txt
discord.py>=2.0,<3.0
aiohttp>=3.9,<4.0
python-dotenv>=1.0
